chore: moves data fetching functions into app/api
