+++ app/tag-data.json
@@ -1,1 +1,1 @@
-{"beginner":11,"career":13,"testing":6,"beginners":1,"webdev":4,"javascript":4,"productivity":10,"discuss":1,"bugs":1,"agile":2,"engineeringnotes":1,"webcomponents":1,"showdev":7,"devops":1,"aws":1,"senior":4,"staff":1,"wishiknew":2,"typescript":2,"angular":2,"webperf":2,"jest":1,"react":4,"teamwork":1,"management":1,"prioritization":1,"email":1,"outlook":1,"git":2,"deploy":1,"nearlyfreespeech":1,"shecoded":4,"css":1,"interview":1,"layoff":1,"mentalhealth":1,"impostersyndrome":1,"growth":1,"mongodb":1,"backend":1,"intermediate":3,"virtualmachine":1,"ie11":1,"responsivedesign":1,"frontend":2,"storybook":1,"leadership":1,"node":1,"heroku":1,"debugging":1,"sql":1,"shecodedally":1,"jobsecurity":1,"enterprise":1}
+{"beginner":11,"career":14,"testing":6,"beginners":1,"webdev":4,"javascript":4,"productivity":11,"discuss":1,"bugs":1,"agile":2,"engineeringnotes":1,"webcomponents":1,"showdev":8,"devops":1,"aws":1,"senior":4,"staff":1,"wishiknew":2,"typescript":2,"angular":2,"webperf":2,"jest":1,"react":4,"teamwork":1,"management":1,"prioritization":1,"email":1,"outlook":1,"git":2,"deploy":1,"nearlyfreespeech":1,"shecoded":4,"css":2,"interview":1,"layoff":1,"mentalhealth":1,"impostersyndrome":1,"growth":1,"mongodb":1,"backend":1,"intermediate":3,"virtualmachine":1,"ie11":1,"responsivedesign":1,"frontend":2,"storybook":1,"leadership":1,"node":1,"heroku":1,"debugging":1,"sql":1,"shecodedally":1,"jobsecurity":1,"enterprise":1,"hottake":1}
